backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
anthropic>=0.40.0
python-multipart>=0.0.12
python-dotenv>=1.0.0
Pillow>=10.2.0
# rembg==2.0.50  # Optional: uncomment if you need background removal

